Move generic error views from `dso_api` => `rest_framework_dso`.
These views exist to conform with the DSO standard, and aren't that
project specific. It also allows better integration with DRF error
handling (next commit).
